+++ js/main.js
@@ -1,7 +1,5 @@
-import { createThumbnails, showBigPicture } from './thumbnails.js';
+import { createThumbnails } from './thumbnails.js';
 import { generatePhotos } from './data.js';
 
 const dummyPhotos = generatePhotos();
 createThumbnails(dummyPhotos);
-
-// showBigPicture(dummyPhotos[0]);
